+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
-- Removed `Architectury API` dependencies
+- try fix compatible with include LDL mods
+- `Forgified Fabric API` compat